requirements.txt
Celery==4.2.1
dj-database-url==0.5.0
Django==2.1
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.6
kombu==4.2.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
requests==2.18.4
whitenoise==4.0
redis==2.10.6
django-redis==4.9.0
